requirements.txt
pytz
pillow
psutil
motor
pymongo[srv]
requests
pyfiglet
aiohttp_requests
disnake
humanize
captcha
httplib2
pycryptodome
onesignal-notifications
disnake-jishaku